+++ lib/cubit/async_cubit.dart
@@ -1,3 +1,6 @@
+import 'dart:developer';
+
+import 'package:warped_bloc/config/default_error_config.dart';
 import 'package:warped_bloc/warped_bloc.dart';
 
 /// This cubit is a wrap around for easily dealing with
@@ -18,7 +21,9 @@ class AsyncCubit extends Cubit<BlocState> {
     } on Failure catch (e) {
       emit(ErrorState(message: e.message));
     } catch (e) {
-      emit(ErrorState(message: e.toString()));
+      final message = await DefaultErrorConfig.onErrorCaught(e);
+      log(e.toString());
+      emit(ErrorState(message: message ?? "Unknown Error has occured"));
     }
   }
 }